renaming show_templatetags to show_template_tags #767
